requirements.txt
black==21.10b0
click==8.1.3
Flask==2.2.3
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.0.3
MarkupSafe==2.1.2
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.0
pygame==2.1.2
python-dotenv==1.0.0
regex==2022.10.31
SQLAlchemy==1.4.29
tomli==1.2.3
typing_extensions==4.4.0
Werkzeug==2.2.3
